+++ react/slider/src/Slider.tsx
@@ -1,14 +1,14 @@
 import { type RefObject, useCallback, useEffect, useRef, useState } from 'react';
 import type { SliderProps } from './props.ts';
 import { clamp, clampValues, roundPrecise, roundIfNumber } from './math.ts';
-import { Tooltip, TooltipArrow } from './Tooltip.tsx';
+import { getTooltipCSS, Tooltip, TooltipArrow } from './Tooltip.tsx';
 import style from 'inline:./style.css';
 
-const thumbWidth = 28;
+export const thumbWidth = 28;
 
 type ObjectRangeValue = { label: string; [key: string]: any };
 type RangeValue = string | ObjectRangeValue;
-type Ref = RefObject<HTMLElement>;
+export type Ref = RefObject<HTMLElement>;
 
 /*
 Slider / range slider component.
@@ -667,106 +667,6 @@ export function Slider({
     </>
   );
 }
-
-// Get full tooltip CSS, to set the position of the tooltips along the slider.
-// This calculates both tooltip box and arrow positions. This is done by calculating the offset from the left,
-// adjusting for input type="range" offset, and adjusting for tooltip box position if containTooltips is used.
-const getTooltipCSS = (
-  currentValues: number[],
-  wrapperRef: Ref,
-  isRange: boolean,
-  max: number,
-  min: number,
-  i: number,
-  widthref: Ref,
-  containTooltips: boolean,
-) => {
-  const width = wrapperRef.current?.clientWidth || 0;
-
-  const left0 = ((currentValues[0] - min) / (max - min)) * width;
-  const left1 = ((currentValues[1] - min) / (max - min)) * width;
-
-  const [offset0, offset1] = getTooltipOffsets(currentValues, max, min);
-
-  const wrapperRect = wrapperRef.current?.getBoundingClientRect();
-  const left = wrapperRect.left;
-
-  // Get distance from left.
-  const l0 = left0 + offset0 + left;
-  const l1 = left1 + offset1 + left;
-
-  const tooltipBox0 = {
-    left: l0 + 'px',
-    transform: 'translateY(-39px) translateX(-50%)',
-  };
-
-  const tooltipBox1 = {
-    left: l1 + 'px',
-    transform: 'translateY(-39px) translateX(-50%)',
-  };
-
-  // If containTooltips is true, the tooltip boxes only move up to the start/end limits.
-  if (containTooltips) {
-    // The following code used in order to estimate (calculate) the width of the tooltip box, with the given value,
-    // before it's rendered.
-    //
-    // This is needed so that an exact position can be determined, without first rendering and then adjusting the
-    // tooltip, which would cause some flickering/stuttering.
-    //
-    // To do this, the value is rendered in the hidden width-check div, the width is then measured, and that value
-    // is used to calculate tooltip position (before it's rendered).
-    const wOffset = 0.5 * (getEstimatedWidth(currentValues[i], widthref) + thumbWidth);
-
-    const right = wrapperRect.right;
-
-    const getStyle = (left?: number) => ({
-      left: left ? left + 'px' : '',
-      transform: 'translateY(-39px)',
-    });
-
-    if (isRange) {
-      if (l0 + wOffset > right) Object.assign(tooltipBox0, getStyle());
-      if (l0 - wOffset < left) Object.assign(tooltipBox0, getStyle(left));
-    }
-
-    if (l1 + wOffset > right) Object.assign(tooltipBox1, getStyle());
-    if (l1 - wOffset < left) Object.assign(tooltipBox1, getStyle(left));
-  }
-
-  return [
-    // Tooltip boxes.
-    tooltipBox0,
-    tooltipBox1,
-    // Tooltip arrows.
-    {
-      left: l0 + 'px',
-      transform: 'translateY(-7.2px) translateX(-50%)',
-    },
-    {
-      left: l1 + 'px',
-      transform: 'translateY(-7.2px) translateX(-50%)',
-    },
-  ];
-};
-
-// Get tooltip offsets, needed to center the tooltip over the thumb (which doesn't follow the active track exactly; see default input type="range" behavior.)
-const getTooltipOffsets = (values: number[], max: number, min: number) => {
-  const tooltipOffset0 = ((values[0] - min) / (max - min)) * thumbWidth;
-  const tooltipOffset1 = ((values[1] - min) / (max - min)) * thumbWidth;
-  
-  return [0.5 * thumbWidth - tooltipOffset0, 0.5 * thumbWidth - tooltipOffset1];
-};
-
-// Determine (estimate) the width of the tooltip box with the given value, using the width-check element.
-const getEstimatedWidth = (val: number, widthRef: Ref) => {
-  const r = widthRef.current;
-
-  if (r) {
-    r.innerText = val.toString();
-
-    return r.clientWidth;
-  }
-};
 
 // Aria label data for the slider.
 // https://www.digitala11y.com/slider-role/.
